+++ src/main/java/com/chargebee/models/enums/Gateway.java
@@ -32,6 +32,7 @@ public enum Gateway {
     ORBITAL,
     MONERIS_US,
     MONERIS,
+    BLUESNAP,
     NOT_APPLICABLE,
     _UNKNOWN; /*Indicates unexpected value for this enum. You can get this when there is a
     java-client version incompatibility. We suggest you to upgrade to the latest version */